+++ src/output/mod.rs
@@ -5,6 +5,7 @@ mod common;
 mod fullname;
 pub mod html5;
 mod html5_serializer;
+mod html5elements;
 #[cfg(feature = "icu")]
 mod icu_normalization;
 mod normalizer;
@@ -16,7 +17,8 @@ mod xml_serializer;
 pub use common::Indentation;
 pub(crate) use fullname::FullnameSerializer;
 pub(crate) use fullname::NamespaceDeclarations;
-pub(crate) use html5_serializer::{Html5Elements, Html5Serializer};
+pub(crate) use html5_serializer::Html5Serializer;
+pub(crate) use html5elements::Html5Elements;
 pub use normalizer::{NoopNormalizer, Normalizer};
 pub(crate) use pretty::Pretty;
 pub use pretty::PrettyOutputToken;